PAOPN-298: remove duplicate code
