+++ Block/Payment/Billet.php
@@ -16,13 +16,7 @@ use Magento\Framework\View\Element\Template\Context;
 use Magento\Checkout\Model\Session as CheckoutSession;
 use Magento\Sales\Api\Data\OrderInterface as Order;
 use Magento\Sales\Api\Data\OrderPaymentInterface as Payment;
-use Pagarme\Core\Kernel\Repositories\OrderRepository;
-use Pagarme\Core\Kernel\ValueObjects\Id\OrderId;
-use Pagarme\Core\Kernel\ValueObjects\Id\SubscriptionId;
-use Pagarme\Core\Recurrence\Repositories\SubscriptionRepository;
-use Pagarme\Pagarme\Concrete\Magento2CoreSetup;
-use Pagarme\Core\Recurrence\Repositories\ChargeRepository as SubscriptionChargeRepository;
-
+use Pagarme\Pagarme\Helper\Payment\Billet as BilletHelper;
 class Billet extends Template
 {
     protected $checkoutSession;
@@ -31,8 +25,7 @@ class Billet extends Template
      * @param Context $context
      * @param CheckoutSession $checkoutSession
      */
-    public function __construct(Context $context, CheckoutSession $checkoutSession)
-    {
+    public function __construct(Context $context, CheckoutSession $checkoutSession){
         $this->checkoutSession = $checkoutSession;
         parent::__construct($context, []);
     }
@@ -72,75 +65,7 @@ class Billet extends Template
      */
     public function getBilletUrl()
     {
-        $method = $this->getPayment()->getMethod();
-
-        if (strpos($method, "pagarme_billet") === false) {
-            return;
-        }
-        $info = $this->getPayment();
-
-        $boletoUrl = $this->getPayment()->getAdditionalInformation('billet_url');
-
-        Magento2CoreSetup::bootstrap();
-        $boletoUrl = $this->getBoletoLinkFromOrder($info);
-
-        if (!$boletoUrl) {
-            $boletoUrl = $this->getBoletoLinkFromSubscription($info);
-        }
-
-        return $boletoUrl;
-    }
-
-    private function getBoletoLinkFromOrder($info)
-    {
-        $lastTransId = $info->getLastTransId();
-        $orderId = null;
-        if ($lastTransId) {
-            $orderId = substr($lastTransId, 0, 19);
-        }
-
-        if (!$orderId) {
-            return null;
-        }
-
-        $orderRepository = new OrderRepository();
-        $order = $orderRepository->findByPagarmeId(new OrderId($orderId));
-
-        if ($order !== null) {
-            $charges = $order->getCharges();
-            foreach ($charges as $charge) {
-                $transaction = $charge->getLastTransaction();
-                $savedBoletoUrl = $transaction->getBoletoUrl();
-                if ($savedBoletoUrl !== null) {
-                    $boletoUrl = $savedBoletoUrl;
-                }
-            }
-        }
-
-        return $boletoUrl;
-    }
-
-    private function getBoletoLinkFromSubscription($info)
-    {
-        $subscriptionRepository = new SubscriptionRepository();
-        $subscription = $subscriptionRepository->findByCode($info->getOrder()->getIncrementId());
-
-        if (!$subscription) {
-            return null;
-        }
-
-        $chargeRepository = new SubscriptionChargeRepository();
-        $subscriptionId =
-            new SubscriptionId(
-                $subscription->getPagarmeId()->getValue()
-            );
-
-        $charge = $chargeRepository->findBySubscriptionId($subscriptionId);
-
-        if (!empty($charge[0])) {
-            return $charge[0]->getBoletoUrl();
-        }
-
-        return null;
+        $billetHelper = new BilletHelper();
+        return $billetHelper->getBilletUrl($this->getPayment());
     }
 }
