+++ Block/Payment/Info/Billet.php
@@ -14,14 +14,11 @@ namespace Pagarme\Pagarme\Block\Payment\Info;
 
 use Magento\Payment\Block\Info;
 use Magento\Framework\DataObject;
-use Pagarme\Core\Kernel\Repositories\OrderRepository;
 use Pagarme\Core\Kernel\Services\OrderService;
 use Pagarme\Core\Kernel\ValueObjects\Id\OrderId;
-use Pagarme\Core\Kernel\ValueObjects\Id\SubscriptionId;
-use Pagarme\Core\Recurrence\Repositories\ChargeRepository as SubscriptionChargeRepository;
-use Pagarme\Core\Recurrence\Repositories\SubscriptionRepository;
 use Pagarme\Pagarme\Concrete\Magento2CoreSetup;
 use Pagarme\Pagarme\Concrete\Magento2PlatformOrderDecorator;
+use Pagarme\Pagarme\Helper\Payment\Billet as BilletHelper;
 
 class Billet extends Info
 {
@@ -47,76 +44,8 @@ class Billet extends Info
 
     public function getBilletUrl()
     {
-        $method = $this->getInfo()->getMethod();
-
-        if (strpos($method, "pagarme_billet") === false) {
-            return;
-        }
-
-        $boletoUrl =  $this->getInfo()->getAdditionalInformation('billet_url');
-
-        Magento2CoreSetup::bootstrap();
-        $info = $this->getInfo();
-
-        $boletoUrl = $this->getBoletoLinkFromOrder($info);
-
-        if (!$boletoUrl) {
-            $boletoUrl = $this->getBoletoLinkFromSubscription($info);
-        }
-
-        return $boletoUrl;
-    }
-
-    private function getBoletoLinkFromOrder($info)
-    {
-        $lastTransId = $info->getLastTransId();
-        $orderId = null;
-        if ($lastTransId) {
-            $orderId = substr($lastTransId, 0, 19);
-        }
-
-        if (!$orderId) {
-            return null;
-        }
-
-        $orderRepository = new OrderRepository();
-        $order = $orderRepository->findByPagarmeId(new OrderId($orderId));
-        $boletoUrl = null;
-
-        if ($order !== null) {
-            $charges = $order->getCharges();
-            foreach ($charges as $charge) {
-                $transaction = $charge->getLastTransaction();
-                $savedBoletoUrl = $transaction->getBoletoUrl();
-                if ($savedBoletoUrl !== null) {
-                    $boletoUrl = $savedBoletoUrl;
-                }
-            }
-        }
-
-        return $boletoUrl;
-    }
-
-    private function getBoletoLinkFromSubscription($info)
-    {
-        $subscriptionRepository = new SubscriptionRepository();
-        $subscription = $subscriptionRepository->findByCode($info->getOrder()->getIncrementId());
-
-        if (!$subscription) {
-            return null;
-        }
-
-        $chargeRepository = new SubscriptionChargeRepository();
-        $subscriptionId =
-            new SubscriptionId(
-                $subscription->getPagarmeId()->getValue()
-            );
-
-        $charge = $chargeRepository->findBySubscriptionId($subscriptionId);
-
-        if (!empty($charge[0])) {
-            return $charge[0]->getBoletoLink();
-        }
+        $billetHelper = new BilletHelper();
+        return $billetHelper->getBilletUrl($this->getInfo());
     }
 
     public function getTitle()
